Major 0.3 release
* Quote sql queries
* Qualify sql queries with table names
* Support internal (db assigned) ids
* Return record after create/update
* Cleaned up doc tests using a macro removing doctest internal feature
  flag
* Consolidated proc macro state in to a single config struct that is
  passed around
