+++ Cargo.toml
@@ -6,7 +6,7 @@ members = [
 
 [package]
 name = "sqlx-crud"
-version = "0.2.0"
+version = "0.3.0"
 edition = "2021"
 readme = "README.md"
 repository = "https://github.com/treydempsey/sqlx-crud"
