+++ sqlx-crud-macros/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "sqlx-crud-macros"
-version = "0.2.0"
+version = "0.3.0"
 edition = "2021"
 repository = "https://github.com/treydempsey/sqlx-crud"
 documentation = "https://docs.rs/sqlx-crud"
